move BaseNode into seperate file and refactor exports
additionally adds stubs for configs
