+++ src/util.rs
@@ -1,4 +1,4 @@
-use crate::error::{Error, Result};
+use crate::{Error, Result};
 use nix::unistd::Uid;
 use std::{
     path::{Path, PathBuf},
